moving into em-components.ts
