+++ src/animation/noodles.ts
@@ -1,4 +1,5 @@
-import { Component, EM } from "../ecs/entity-manager.js";
+import { EM } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import {
   cloneMesh,
   mapMeshPositions,
