+++ src/cannons/bullet.ts
@@ -1,4 +1,5 @@
-import { EM, Component, Entity, EntityW } from "../ecs/entity-manager.js";
+import { EM, Entity, EntityW } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import { V2, V3, V4, quat, mat4, V, tV } from "../matrix/sprig-matrix.js";
 import { FinishedDef } from "../ecs/em-helpers.js";
 import { ColorDef } from "../color/color-ecs.js";
