+++ src/ecs/em-resources.ts
@@ -5,7 +5,8 @@ import {
 import { assert, getCallStack } from "../utils/util-no-import.js";
 import { Intersect } from "../utils/util.js";
 import { _init } from "./em-init.js";
-import { componentsToString, nameToId } from "./entity-manager.js";
+import { nameToId } from "./entity-manager.js";
+import { componentsToString } from "./em-components.js";
 
 type ResourcesPromise<RS extends ResourceDef[]> = {
   id: number;
