+++ src/ecs/entity-pool.ts
@@ -1,4 +1,5 @@
-import { ComponentDef, Entity, EntityW } from "./entity-manager.js";
+import { Entity, EntityW } from "./entity-manager.js";
+import { ComponentDef } from "./em-components.js";
 import { IdxPool, createIdxPool } from "../utils/idx-pool.js";
 import { assert } from "../utils/util.js";
 import { never } from "../utils/util-no-import.js";
