+++ src/input/interact.ts
@@ -1,4 +1,5 @@
-import { Component, EM, Entity } from "../ecs/entity-manager.js";
+import { EM, Entity } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import { LocalPlayerEntityDef } from "../hyperspace/hs-player.js";
 import { V2, V3, V4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import { AuthorityDef, MeDef } from "../net/components.js";
