+++ src/ld53/ship-health.ts
@@ -1,4 +1,5 @@
-import { Component, EM } from "../ecs/entity-manager.js";
+import { EM } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import { WoodHealthDef } from "../wood/wood.js";
 import { Phase } from "../ecs/sys-phase.js";
