+++ src/motion/velocity.ts
@@ -1,4 +1,5 @@
-import { EM, Component } from "../ecs/entity-manager.js";
+import { EM } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import { V2, V3, V4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import { TimeDef } from "../time/time.js";
 import { PositionDef, RotationDef } from "../physics/transform.js";
