+++ src/physics/nonintersection.ts
@@ -1,5 +1,6 @@
 import { Collider, ColliderDef, DefaultLayer, Layer } from "./collider.js";
-import { Component, EM, Entity } from "../ecs/entity-manager.js";
+import { EM, Entity } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import { V2, V3, V4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import {
   CollidesWith,
