+++ src/physics/transform.ts
@@ -1,4 +1,5 @@
-import { Component, EM } from "../ecs/entity-manager.js";
+import { EM } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import { V2, V3, V4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import { WorldFrameDef, doesOverlap } from "./nonintersection.js";
 import { FALSE, dbgLogOnce } from "../utils/util.js";
