+++ src/render/gpu-registry.ts
@@ -1,4 +1,4 @@
-import { ComponentDef } from "../ecs/entity-manager.js";
+import { ComponentDef } from "../ecs/em-components.js";
 import { V2, V3, V4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import { assert } from "../utils/util.js";
 import {
