+++ src/render/mesh-pool.ts
@@ -21,7 +21,7 @@ import {
 } from "./gpu-registry.js";
 import { V2, V3, V4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import { DEFAULT_MASK } from "./pipeline-masks.js";
-import { ComponentDef } from "../ecs/entity-manager.js";
+import { ComponentDef } from "../ecs/em-components.js";
 import { GPUBufferUsage } from "./webgpu-hacks.js";
 import { CyResources } from "./instantiator-webgpu.js";
