+++ src/render/motion-smoothing.ts
@@ -1,4 +1,5 @@
-import { EM, Component, Entity } from "../ecs/entity-manager.js";
+import { EM, Entity } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import { V2, V3, V4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import { TimeDef } from "../time/time.js";
 import {
