+++ src/stone/stone.ts
@@ -4,7 +4,8 @@ import { ColorDef } from "../color/color-ecs.js";
 import { ENDESGA16 } from "../color/palettes.js";
 import { DeadDef } from "../ecs/delete.js";
 import { createRef, Ref } from "../ecs/em-helpers.js";
-import { Component, EM, Entity, EntityW } from "../ecs/entity-manager.js";
+import { EM, Entity, EntityW } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import { EntityPool, createEntityPool } from "../ecs/entity-pool.js";
 import { Bullet, BulletDef, fireBullet } from "../cannons/bullet.js";
 import { GravityDef } from "../motion/gravity.js";
