+++ src/turret/yawpitch.ts
@@ -1,5 +1,6 @@
 import { defineSerializableComponent } from "../ecs/em-helpers.js";
-import { EM, Component } from "../ecs/entity-manager.js";
+import { EM } from "../ecs/entity-manager.js";
+import { Component } from "../ecs/em-components.js";
 import { V2, V3, V4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 
 // TODO(@darzu): do we really need this component? Ideally we'd just have helpers around
